add user.rdo
