+++ src/shared/libs/modules/users/rdo/user.rdo.ts
@@ -1,0 +1,15 @@
+import { Expose } from 'class-transformer';
+
+export class UserRdo {
+  @Expose()
+  public email: string ;
+
+  @Expose()
+  public avatarPath: string;
+
+  @Expose()
+  public firstname: string;
+
+  @Expose()
+  public lastname: string;
+}
