+++ src/shared/libs/rest/controller/exception-filter/app-exception-filter.ts
@@ -1,0 +1,39 @@
+import { Request, Response, NextFunction } from 'express';
+import { ExceptionFilter } from './exception-filter.interface.js';
+import { inject, injectable } from 'inversify';
+import { StatusCodes } from 'http-status-codes';
+import { Logger } from '../../../logger/index.js';
+import { Component } from '../../../../types/component.enum.js';
+import { createErrorObject } from '../../../../helpers/index.js';
+import { HttpError } from '../../../rest/errors/index.js';
+
+@injectable()
+export class AppExceptionFilter implements ExceptionFilter {
+  constructor(
+    @inject(Component.Logger) private readonly logger: Logger
+  ){
+    this.logger = logger;
+  }
+
+  private handleHttpError(error: HttpError, _req: Request, res: Response, _next: NextFunction) {
+    this.logger.error(`[${error.detail}]: ${error.httpStatusCode} — ${error.message}`, error);
+    res
+      .status(error.httpStatusCode)
+      .json(createErrorObject(error.message));
+  }
+
+  private handleOtherError(error: Error, _req: Request, res: Response, _next: NextFunction) {
+    this.logger.error(error.message, error);
+    res
+      .status(StatusCodes.INTERNAL_SERVER_ERROR)
+      .json(createErrorObject(error.message));
+  }
+
+  public catch(error: Error | HttpError, req: Request, res: Response, next: NextFunction): void {
+    if (error instanceof HttpError) {
+      return this.handleHttpError(error, req, res, next);
+    }
+
+    this.handleOtherError(error, req, res, next);
+  }
+}
