--- src/shared/libs/rest/controller/exception-filter/app-exception.ts
@@ -1,23 +1,0 @@
-import { Request, Response, NextFunction } from 'express';
-import { ExceptionFilter } from './exception-filter.interface.js';
-import { inject, injectable } from 'inversify';
-import { StatusCodes } from 'http-status-codes';
-import { Logger } from '../../../../libs/logger/index.js';
-import { Component } from '../../../../types/component.enum.js';
-
-@injectable()
-export class AppExceptionFilter implements ExceptionFilter {
-  constructor(
-    @inject(Component.Logger) private readonly logger: Logger
-  ){
-    this.logger = logger;
-  }
-
-  public catch(error: Error, _req: Request, res: Response, _next: NextFunction): void {
-    this.logger.error(error.message, error);
-    res
-      .status(StatusCodes.INTERNAL_SERVER_ERROR)
-      .json({ error: error.message});
-  }
-
-}
