+++ src/shared/libs/rest/controller/exception-filter/index.ts
@@ -1,2 +1,2 @@
-export * from './app-exception.js';
+export * from './app-exception-filter.js';
 export * from './exception-filter.interface.js';
