+++ src/shared/libs/rest/errors/index.ts
@@ -1,0 +1,1 @@
+export * from './http-error.js';
